requirements-macos.txt
absl-py==1.0.0
astunparse==1.6.3
beautifultable==1.0.1
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cloudpickle==1.6.0
cycler==0.11.0
Cython==0.29.27
fasteners==0.17.3
flatbuffers==2.0
fonttools==4.29.1
gast==0.4.0
glfw==2.5.0
glob2==0.7
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1639715812589/work
gym==0.19.0
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964045648/work
idna==3.3
imageio==2.14.1
importlib-metadata==4.10.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
mujoco-py==2.1.2.14
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1640083125610/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pickle5==0.0.11
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stable-baselines3==1.4.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-macos==2.7.0
tensorflow-metal==0.3.0
termcolor==1.1.0
torch==1.9.0
torchaudio==0.11.0
torchvision==0.10.0
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
